requirements.txt
supabase==1.0.3
python-dotenv==1.0.0
reportlab==4.0.4
Pillow==10.0.0
matplotlib==3.7.2
pandas==2.0.3
numpy==1.24.3
requests==2.31.0
python-dateutil==2.8.2 